backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.0.0
aiofiles==0.23.0
jinja2==3.1.2